refactor: split agent.rs into agent/ module — dismiss + tests submodules (#1421)
agent.rs had grown to 3239 LOC (~1700 production + ~1536 tests).
Convert it to a directory module and extract two cohesive units:

- agent/dismiss.rs: the dialog auto-dismiss subsystem (regex cache,
  literal-hint helper, try_dismiss_dialog) + its co-located tests.
  All internal helpers stay private; only try_dismiss_dialog is
  re-exported from mod.rs so external call sites are unchanged.
- agent/tests.rs: the remaining (non-dismiss) test module, declared
  via `#[cfg(test)] #[path] mod tests;` so the spawn-rationale audit's
  test-only-file detection keeps skipping it.

mod.rs retains the core types, registry, spawn/lifecycle/exit, and
PTY-write code — now 1533 LOC. Registry/lifecycle were left in place:
they are interwoven with the core handle types and splitting them
would force broad pub(crate) exposure for little gain. Type-only
move; no behavior change.


Agend-Agent: fixup-dev
Agend-Task: t-20260528154227387386-20
Agend-Branch: refactor/1421-agent-split
Agend-Issued-At: 2026-05-28T15:42:40.883872+00:00
